refactor: extract shared utilities and fix remaining review issues
- Add OS translation keys to French common.json
- Extract debounce utility to shared lib/debounce.ts
  - Returns {fn, cancel} object for proper cleanup
  - Used by both useXterm.ts and LivePreviewTerminal.tsx
- Extract terminal theme to shared lib/terminal-theme.ts
  - DEFAULT_TERMINAL_THEME constant with all 17 color properties
  - Used by both useXterm.ts and LivePreviewTerminal.tsx
- Add cancel cleanup to debounce in useXterm.ts
  - Prevents pending debounced calls from firing after unmount
  - Fixes React warning about setState on unmounted component

These changes address CodeRabbit review feedback about:
- Missing French translations for OS names
- Duplicated debounce utility functions
- Duplicated terminal theme configuration
- Missing cleanup for debounced resize timeout
